Add a texture data based silhouette shader
Now that we have a texture-data light shader, we need a corresponding
silhouette version. That's not too difficult - it's mostly just filing
off the serial numbers, if you will.

It works, but I'm not very happy with it. It seems as though the
lights aren't reaching upward as they should be, though maybe that's
how it's supposed to be? I'm not sure.

I think this gonna require a rework. A crazy, insane, particle-driven
rework...