Add rooftop decor shaders
The rooftop decorations are currently unlit, so I had the idea of
using particles to spawn them. We would then use the particle's
emitter position to index our light data. That would allow the
decorations to be offset from the building's center while still using
the same lights.

So I added the shaders necessary to do this. I had to readapt the
Particle Shader from one a base Particle Shader converted to source
code.